Changed file names for guess number game
